+++ .gitignore
@@ -2,6 +2,7 @@
 /target/
 **/*.rs.bk
 *.pdb
+/src/ui.rs.backup
 
 # C
 *.o
