+++ src/ui/screens.rs
@@ -1,0 +1,30 @@
+#[derive(Clone, PartialEq)]
+pub enum Screen {
+    VaultCheck,
+    CreateVault,
+    UnlockVault,
+    MainMenu,
+    ViewPasswords,
+    AddPassword,
+    EditPassword,
+    ViewHistory,
+    SearchPassword,
+    GeneratePassword,
+    DeletePassword,
+    FilterByTag,
+}
+
+#[derive(Clone, PartialEq)]
+pub enum InputField {
+    None,
+    Password,
+    PasswordConfirm,
+}
+
+#[derive(Clone, PartialEq)]
+pub enum MessageType {
+    None,
+    Success,
+    Error,
+    Info,
+}
